Broke the localization tests to their own file.
Put shared test functions into a shared file. Pulled test function to all callers of the test functions, because the single-file page cannot be passed through to another file's tests.
