+++ tests/projects/main/sections/sections.spec.types.ts
@@ -1,0 +1,8 @@
+import type { SheetHelperItemCreationArgs } from 'tests/helpers/SheetHelper';
+
+export type DefaultSectionTestParams = {
+  itemCreationArgs: SheetHelperItemCreationArgs;
+  tabId: string;
+  sectionKey: string;
+  sectionLabel?: string;
+};
